VDP2: Rework color fetching for VDP1 data

+++ libs/satemu-core/include/satemu/hw/vdp/vdp_defs.hpp
@@ -784,7 +784,7 @@ struct SpriteParams {
 
     // Sprite color data offset.
     // Derived from CRAOFB.SPCAOSn
-    uint16 colorDataOffset;
+    uint32 colorDataOffset;
 };
 
 struct SpriteData {
